--- a/src/Gas.sol
+++ b/src/Gas.sol
@@ -1,15 +1,13 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.0; 
 
-import "./Ownable.sol";
-
 contract Constants {
     uint256 public constant tradeFlag = 1;
     uint256 public constant basicFlag = 0;
     uint256 public constant dividendFlag = 1;
 }
 
-contract GasContract is Ownable, Constants {
+contract GasContract is Constants {
     uint256 public totalSupply = 0; // cannot be updated
     uint256 public paymentCounter = 0;
     mapping(address => uint256) public balances;