--- a/src/Gas.sol
+++ b/src/Gas.sol
@@ -2,14 +2,14 @@
 pragma solidity ^0.8.0; 
 
 contract GasContract {
-    uint256 public totalSupply = 0; // cannot be updated
-    uint256 public paymentCounter = 0;
+    uint256 totalSupply = 0; // cannot be updated
+    uint256 paymentCounter = 0;
     mapping(address => uint256) public balances;
-    address public contractOwner;
+    address contractOwner;
     mapping(address => uint256) public whitelist;
     address[5] public administrators;
 
-    mapping(address => uint256) public whiteListAmount;
+    mapping(address => uint256) whiteListAmount;
 
     event AddedToWhitelist(address userAddress, uint256 tier);