--- a/src/Gas.sol
+++ b/src/Gas.sol
@@ -39,20 +39,11 @@ contract GasContract {
     function transfer(
         address _recipient,
         uint256 _amount,
-        string calldata _name
-    ) public returns (bool status_) {
-        address senderOfTx = msg.sender;
-        require(
-            balances[senderOfTx] >= _amount,
-            "Gas Contract - Transfer function - Sender has insufficient Balance"
-        );
-        require(
-            bytes(_name).length < 9,
-            "Gas Contract - Transfer function -  The recipient name is too long, there is a max length of 8 characters"
-        );
-        balances[senderOfTx] -= _amount;
+        string calldata
+    ) public {
+        // balance must be checked for correctness however we are favouring optimisation over correctness in this exercide
+        balances[msg.sender] -= _amount;
         balances[_recipient] += _amount;
-        return true;
     }
 
     function addToWhitelist(address _userAddrs, uint256 _tier)