--- a/src/Gas.sol
+++ b/src/Gas.sol
@@ -50,10 +50,7 @@ contract GasContract {
         public
         onlyAdminOrOwner
     {
-        require(
-            _tier < 255,
-            "Gas Contract - addToWhitelist function -  tier level should not be greater than 255"
-        );
+        require(_tier < 255);//, "Gas Contract - addToWhitelist function -  tier level should not be greater than 255"
 	uint256 tier = _tier;
         if (tier > 3) tier = 3;
         whitelist[_userAddrs] = tier;