Gas optimisation - commit history shows work step by step.
```
% forge test --gas-report
[⠊] Compiling...
[⠑] Compiling 25 files with Solc 0.8.28
[⠘] Solc 0.8.28 finished in 1.74s
Compiler run successful!

Ran 12 tests for test/Gas.UnitTests.t.sol:GasTest
[PASS] testAddToWhitelist(uint256) (runs: 256, μ: 35328, ~: 35262)
[PASS] testBalanceOf() (gas: 15125)
[PASS] testCheckForAdmin() (gas: 11239)
[PASS] testTransfer(uint256,address) (runs: 256, μ: 88682, ~: 89562)
[PASS] testWhiteTranferAmountUpdate(uint256,string,uint256) (runs: 256, μ: 186385, ~: 186452)
[PASS] test_admins() (gas: 28728)
[PASS] test_onlyOwner(address,uint256) (runs: 256, μ: 34795, ~: 34897)
[PASS] test_tiers(address,uint256) (runs: 256, μ: 38001, ~: 38113)
[PASS] test_tiersReverts(address,uint256) (runs: 256, μ: 34237, ~: 34255)
[PASS] test_whitelistEvents(address,address,uint256,string,uint256) (runs: 256, μ: 177230, ~: 177351)
[PASS] test_whitelistEvents(address,uint256) (runs: 256, μ: 40942, ~: 41051)
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 177318, ~: 177614)
Suite result: ok. 12 passed; 0 failed; 0 skipped; finished in 443.12ms (2.08s CPU time)

╭----------------------------------+-----------------+-------+--------+-------+---------╮
| src/Gas.sol:GasContract Contract |                 |       |        |       |         |
+=======================================================================================+
| Deployment Cost                  | Deployment Size |       |        |       |         |
|----------------------------------+-----------------+-------+--------+-------+---------|
| 257431                           | 1566            |       |        |       |         |
|----------------------------------+-----------------+-------+--------+-------+---------|
|                                  |                 |       |        |       |         |
|----------------------------------+-----------------+-------+--------+-------+---------|
| Function Name                    | Min             | Avg   | Median | Max   | # Calls |
|----------------------------------+-----------------+-------+--------+-------+---------|
| addToWhitelist                   | 21566           | 22600 | 22916  | 23132 | 2048    |
|----------------------------------+-----------------+-------+--------+-------+---------|
| administrators                   | 387             | 424   | 431    | 448   | 5       |
|----------------------------------+-----------------+-------+--------+-------+---------|
| balanceOf                        | 514             | 1955  | 2532   | 2532  | 1793    |
|----------------------------------+-----------------+-------+--------+-------+---------|
| balances                         | 470             | 970   | 470    | 2470  | 1024    |
|----------------------------------+-----------------+-------+--------+-------+---------|
| checkForAdmin                    | 383             | 383   | 383    | 383   | 1       |
|----------------------------------+-----------------+-------+--------+-------+---------|
| getPaymentStatus                 | 438             | 438   | 438    | 438   | 256     |
|----------------------------------+-----------------+-------+--------+-------+---------|
| transfer                         | 26994           | 66495 | 66794  | 66890 | 1024    |
|----------------------------------+-----------------+-------+--------+-------+---------|
| whiteTransfer                    | 67462           | 67661 | 67714  | 67726 | 768     |
|----------------------------------+-----------------+-------+--------+-------+---------|
| whitelist                        | 338             | 338   | 338    | 338   | 512     |
╰----------------------------------+-----------------+-------+--------+-------+---------╯


Ran 1 test suite in 515.58ms (443.12ms CPU time): 12 tests passed, 0 failed, 0 skipped (12 total tests)
```